# openapi.yaml
openapi: 3.0.3
info:
  title: Payment Gateway API
  version: 1.0.0
  description: |
    Stripe-style REST API for the in-house payment gateway.

    ## Authentication

    Two parallel auth systems:

    | Use case | Header |
    |----------|--------|
    | Merchant API consumers | `Authorization: Bearer sk_test_xxx_yyy` (or `sk_live_…`) |
    | Dashboard sessions | `Authorization: Bearer <jwt>` (issued by `/auth/login` + `/auth/totp/verify`) |

    Public keys (`pk_test_…` / `pk_live_…`) are restricted to the `tokenize` scope and are safe to ship to the browser.

    ## Idempotency

    Every `POST /v1/payment_intents` request **must** include an `Idempotency-Key` header.
    Replaying with the same `(merchantId, idempotencyKey)` returns the original payment intent.

    ## Pagination

    List endpoints use cursor-based pagination:

    ```
    GET /v1/customers?limit=20&cursor=cus_abc123
    → { data: [...], has_more: true, next_cursor: "cus_xyz789" }
    ```

    ## Errors

    All errors follow a consistent shape:

    ```json
    { "error": { "code": "validation_error", "message": "...", "param": "amount" } }
    ```
  contact:
    name: Payments Engineering
    email: payments@example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://sandbox.payments.example.com/api/v1
    description: Sandbox
  - url: https://api.payments.example.com/api/v1
    description: Production
...
components:
  securitySchemes:
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: api_key
      description: |
        Merchant API key. Format: `sk_test_xxx_yyy` or `sk_live_xxx_yyy`.
        Public keys (`pk_…`) are restricted to the `tokenize` scope.
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Dashboard session JWT issued by `POST /auth/login` (partial)
        and elevated by `POST /auth/totp/verify` (full).
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorParam:
      name: cursor
      in: query
      schema:
        type: string
      description: ID of the last item in the previous page
  schemas:
...
    Webhook:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        events:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [ACTIVE, DISABLED, FAILING]
        createdAt: { type: string, format: date-time }
...
paths:
...
  /webhooks:
    get:
      tags: [Webhooks]
      summary: List webhooks for the merchant
      security: [{ jwtAuth: [] }]
      responses:
        '200':
          description: Webhook list (with delivery counts).
    post:
      tags: [Webhooks]
      summary: Register a new webhook endpoint
      description: |
        Returns the **plaintext** signing secret exactly once.
        Verify incoming events using the `X-Payments-Signature` header — see
        [signature spec](#section/Webhook-signatures).
      security: [{ jwtAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url: { type: string, format: uri }
                events:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Webhook registered, secret returned once.
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhook: { $ref: '#/components/schemas/Webhook' }
                  secret: { type: string, example: whsec_xxx }
                  note: { type: string }
  /webhooks/{id}:
    patch:
      tags: [Webhooks]
      summary: Update a webhook (URL, events, status)
      security: [{ jwtAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string, format: uri }
                events:
                  type: array
                  items: { type: string }
                status:
                  type: string
                  enum: [ACTIVE, DISABLED]
      responses:
        '200':
          description: Updated.
    delete:
      tags: [Webhooks]
      summary: Delete a webhook
      security: [{ jwtAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Deleted.
...
  /webhooks/{id}/test:
    post:
      tags: [Webhooks]
      summary: Send a synthetic webhook.test event to this endpoint
      security: [{ jwtAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Test event enqueued for delivery.
  /events:
    get:
      tags: [Events]
      summary: Replay / inspect events for the merchant
      security: [{ jwtAuth: [] }]
      parameters:
        - { name: type, in: query, schema: { type: string }, description: 'Exact event type filter, e.g. payment_intent.succeeded' }
        - { name: since, in: query, schema: { type: string, format: date-time } }
        - { name: until, in: query, schema: { type: string, format: date-time } }
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Cursor-paginated events.
  /events/{id}:
    get:
      tags: [Events]
      summary: Get an event by ID, with its delivery attempts
      security: [{ jwtAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Event detail with deliveries array.
...
security:
  - apiKeyAuth: []